requirements.txt
Flask==3.0.0
RPi.GPIO==0.7.1
python-dotenv==1.0.1